refactor(webviews): consume @microsoft/vscode-webview-api
Replaces the in-tree src/webviews/api/{extension-server,webview-client,
configuration}/ folders with thin app-specific wrappers that bind the
generic framework package to DocumentDB:

- src/webviews/api/appRouter.ts        — root tRPC tree +
  DocumentDB BaseRouterContext (extends the framework base with
  dbExperience + webviewName; telemetry typed against the framework's
  TelemetryContext)
- src/webviews/api/WebviewRegistry.ts  — app-specific webview keys
- src/webviews/api/trpc.ts             — DocumentDB telemetry
  middleware preserving the existing documentDB.rpc.* event-name
  convention; redefines WithTelemetry<T> to surface ITelemetryContext
  (suppressAll, etc.)
- src/webviews/api/WebviewController.ts — DocumentDB-tuned base class
  pre-filling the framework's WebviewControllerOptions with appRouter,
  bundle layout, and dev-server host
- src/webviews/api/useTrpcClient.ts    — AppRouter-bound hook wrapper

Webview components import accessibility/configuration helpers directly
from @microsoft/vscode-webview-api; extension-host code imports from
@microsoft/vscode-webview-api/server. The old WebviewContext.tsx and
the three subfolders under src/webviews/api/ are removed.
