--- src/webviews/api/webview-client/accessibility/index.ts
@@ -1,7 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-export { Announcer } from './Announcer';
-export type { AnnouncerProps } from './Announcer';
