+++ src/webviews/documentdb/collectionView/components/queryInsightsTab/components/optimizationCards/custom/GetPerformanceInsightsCard.tsx
@@ -15,9 +15,9 @@ import {
     tokens,
 } from '@fluentui/react-components';
 import { SparkleRegular } from '@fluentui/react-icons';
+import { Announcer } from '@microsoft/vscode-webview-api';
 import * as l10n from '@vscode/l10n';
 import type * as React from 'react';
-import { Announcer } from '../../../../../../../api/webview-client/accessibility';
 import '../baseOptimizationCard.scss';
 import './GetPerformanceInsightsCard.scss';
