+++ src/webviews/documentdb/documentView/documentsViewController.ts
@@ -6,7 +6,7 @@
 import * as vscode from 'vscode';
 import { API } from '../../../DocumentDBExperiences';
 import { ext } from '../../../extensionVariables';
-import { WebviewController } from '../../api/extension-server/WebviewController';
+import { WebviewController } from '../../api/WebviewController';
 import { type RouterContext } from './documentsViewRouter';
 
 export type DocumentsViewWebviewConfigurationType = {
